refactor: remove code specific to ArkTS architecture and improve names
See merge request rnoh/react-native-harmony-gesture-handler!81

(cherry picked from commit 2af6c2753d161639dfe7771e3198a8231c211952)
