+++ tester/harmony/gesture_handler/index.ets
@@ -1,4 +1,2 @@
 export * from "./ts"
-export * from "./src/main/ets/rnoh/RNGestureHandlerRootView"
-export * from "./src/main/ets/rnoh/RNGestureHandlerButton"
 export { default } from "./src/main/ets/RNOHPackage"
