--- tester/harmony/gesture_handler/src/main/ets/core/ViewFinder.ts
@@ -1,11 +1,0 @@
-import { Tag, View } from "./View"
-
-export interface ViewFinder {
-    getTouchableViewsAt(
-      pointRelativeToRoot: {
-        x: number,
-        y: number
-      },
-      rootTag: Tag
-    ): View[]
-  }
