+++ tester/harmony/gesture_handler/src/main/ets/rnoh/RNOHScrollLocker.ts
@@ -1,14 +1,7 @@
 import { RNInstance } from '@rnoh/react-native-openharmony/ts';
 import { ScrollLocker, RNGHLogger } from '../core';
 
-export class RNOHScrollLockerArkTS implements ScrollLocker {
-  constructor(private rnInstance: RNInstance) {
-  }
 
-  lockScrollContainingViewTag(viewTag: number) {
-    return this.rnInstance.blockComponentsGestures(viewTag);
-  }
-}
 
 export class RNOHScrollLockerCAPI implements ScrollLocker {
   private logger: RNGHLogger
